% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_final.R
\name{add_diff_predictors}
\alias{add_diff_predictors}
\title{add_diff_predictors}
\usage{
add_diff_predictors(obj, categories = list())
}
\arguments{
\item{obj}{wizard object}

\item{categories}{List of categories to generate the iterated difference feature}
}
\value{
wizard object
}
\description{
Function to generate iterated difference between lagged features
}
