% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_final.R
\name{add_lagged_predictors}
\alias{add_lagged_predictors}
\title{add_lagged_predictors}
\usage{
add_lagged_predictors(
  obj,
  lookback,
  window_size,
  step,
  lookahead,
  feature_stat,
  impute
)
}
\arguments{
\item{obj}{wizard object}

\item{lookback}{lookback period for each category}

\item{window_size}{window_size for each category}

\item{step}{Number of windows to move forward}

\item{lookahead}{The outcome period}

\item{feature_stat}{The summary statistics to compute for each category}

\item{impute}{T/F to carry forward impute}
}
\value{
wizard object
}
\description{
The function generates the lagged features for every window
}
