% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_final.R
\name{add_outcome}
\alias{add_outcome}
\title{add_outcome}
\usage{
add_outcome(obj, outcome_var, outcome_stat = list())
}
\arguments{
\item{obj}{wizard object}

\item{outcome_var}{outcome variable}

\item{outcome_stat}{outcome statistics}
}
\value{
wizard object
}
\description{
The function to generate the outcome stat for the data
}
