% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_final.R
\name{add_prop_predictors}
\alias{add_prop_predictors}
\title{add_prop_predictors}
\usage{
add_prop_predictors(obj, categories = list())
}
\arguments{
\item{obj}{wizard object}

\item{categories}{List of categories to generate teh proportional variation feature}
}
\value{
wizard object
}
\description{
Function to generate proportional variation between lagged features
}
