% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_final.R
\name{build_wizard_object}
\alias{build_wizard_object}
\title{Build wizard object}
\usage{
build_wizard_object(temporal_data, fixed_data)
}
\arguments{
\item{temporal_data}{temporal dataframe}

\item{fixed_data}{fixed data}
}
\value{
wizard object
}
\description{
Creates and initialized a wizard object with temporal and fixed data frame
}
