% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_col_names_generator.R
\name{categorical_col_names_generator}
\alias{categorical_col_names_generator}
\title{Output file column names generator}
\usage{
categorical_col_names_generator(temporal_data)
}
\arguments{
\item{temporal_data}{Data frame}
}
\value{
A matrix of the infile
}
\description{
The colnames_generator function is a  function to identify categorical features and combine them with their values and provide a true value for their presence in datafile.
}
