% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapper.R
\name{check_mapper}
\alias{check_mapper}
\title{Check mapper function}
\usage{
check_mapper(
  .x,
  .y,
  window_size = list(meds = 1, labs = 6),
  step = 3,
  feature_stat = list(labs = c("min", "mean", "max"), meds = ("min"))
)
}
\arguments{
\item{window_size}{window size for each category}

\item{step}{step size}

\item{feature_stat}{stats to compute for each feature}
}
\description{
Check mapper function
}
