% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames_generator.R
\name{colnames_generator}
\alias{colnames_generator}
\title{Creates features}
\usage{
colnames_generator(temporal_data, feature_stat = list(labs = c("min",
  "mean", "max", "n"), meds = ("n")), lagged_feature_stat = list(labs =
  c("min", "mean")), default_stats = c("min", "max", "mean", "n"),
  lookback = 48, window_size = 6)
}
\arguments{
\item{temporal_data}{Data frame}

\item{feature_stat}{List of statistical methods to use}

\item{lagged_feature_stat}{feature stats}

\item{default_stats}{default stats}

\item{lookback}{lookback period}

\item{window_size}{window size}
}
\value{
Variable list
}
\description{
The function is intended to create a expanded list of all the features to be created after transforming the data in way.
}
