% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wisard.R
\name{create_wisard}
\alias{create_wisard}
\title{Create wisard}
\usage{
create_wisard(file = "", fixed_data = NULL, window_size = 6,
  lookback = 48, lookahead = 30, feature_stat = list(labs = c("min",
  "mean", "max", "n"), meds = ("n")), outcome_var = NULL,
  outcome_stat = list("mean", "min", ",max"), lag = FALSE,
  impute = T, write_file = NULL)
}
\arguments{
\item{file}{temporal data file}

\item{fixed_data}{fixed data file}

\item{window_size}{size of window: numeric when time var is numeric or period (e.g. hour(6)) when time is date object}

\item{lookback}{lookback period}

\item{lookahead}{lookahead period}

\item{feature_stat}{feature stats}

\item{outcome_var}{outcome variable}

\item{outcome_stat}{outcome stats}

\item{lag}{whether to include lag features}

\item{impute}{whether to impute data}

\item{write_file}{path to csv output}
}
\description{
Create wisard
}
