% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_time.R
\name{date_to_time}
\alias{date_to_time}
\title{Date to time}
\usage{
date_to_time(temporal_data = temporal_data, fixed_data = fixed_data,
  units)
}
\arguments{
\item{temporal_data}{temporal df}

\item{units}{units}

\item{fixed_datafixed}{df}
}
\value{
data frame
}
\description{
Convert the timestamps in the temporal data to Number relative to the admit time or initial time.
}
