% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_feature.R
\name{diff_feature}
\alias{diff_feature}
\title{Difference features}
\usage{
diff_feature(final_frame, window_size)
}
\arguments{
\item{final_frame}{final df}

\item{window_size}{window size}
}
\value{
lag frame
}
\description{
The function to compute the difference features.
}
