% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_outcome_variable.R
\name{dummy_outcome_variable}
\alias{dummy_outcome_variable}
\title{Dummy outcome variable}
\usage{
dummy_outcome_variable(temporal_data, outcome_var = NULL,
  outcome_stat = list("mean"), lookahead = 30, window_size = 6)
}
\arguments{
\item{temporal_data}{temporal df}

\item{outcome_var}{outcome variable}

\item{outcome_stat}{outcome statistics}

\item{lookahead}{lookahead}

\item{window_size}{window size}
}
\value{
lag frame
}
\description{
The function does not spread the outcome variable to
}
