% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_wisard.R
\name{dummy_wisard}
\alias{dummy_wisard}
\title{Dummy Wisard}
\usage{
dummy_wisard(temporal_data,
  all_variable_to_create = all_variable_to_create,
  lag_variable_to_create = lag_variable_to_create, window_size = 6,
  outcome_var = NULL, outcome_stat = list("mean", "min", "max"),
  lookahead = 30, lookback = 48, lag = FALSE, impute = T)
}
\arguments{
\item{temporal_data}{temporal df}

\item{all_variable_to_create}{vars to create}

\item{lag_variable_to_create}{lag vars to create}

\item{window_size}{window size}

\item{outcome_var}{outcome variable}

\item{outcome_stat}{outcome stats}

\item{lookahead}{lookahead}

\item{lookback}{lookback}

\item{lag}{T/F create lags}

\item{impute}{T/F impute}
}
\value{
final_frame
}
\description{
Dummy Wisard
}
