% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hour_to_number.R
\name{hour_to_number}
\alias{hour_to_number}
\title{Hour to number}
\usage{
hour_to_number(
  window_size = NULL,
  lookback = NULL,
  lookahead = NULL,
  step = NULL
)
}
\arguments{
\item{window_size}{window size}

\item{lookback}{lookback}

\item{lookahead}{lookahead}

\item{step}{step}
}
\value{
list of values
}
\description{
convert the user argument which is a period object to a numeric value
}
