% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapper.R
\name{iterative_lag_features}
\alias{iterative_lag_features}
\title{iterative_lag_features}
\usage{
iterative_lag_features(final_frame, categories, window_size, lag_compute)
}
\arguments{
\item{final_frame}{data frame of lagged features}

\item{categories}{categories}

\item{window_size}{window size}
}
\description{
iterative_lag_features
}
