% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope.R
\name{slope}
\alias{slope}
\title{Slope}
\usage{
slope(time, value)
}
\arguments{
\item{time}{time}

\item{value}{value}
}
\value{
slope value
}
\description{
Designing a function to compute the slope of the most recent window.
}
