% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapper.R
\name{step_lag}
\alias{step_lag}
\title{Step lag function is a internal function to adjust to the step value.}
\usage{
step_lag(temporal_data, step, window_size, category)
}
\arguments{
\item{temporal_data}{temporal data frame}

\item{step}{step size}

\item{window_size}{window size}

\item{category}{category of the feature}
}
\description{
Step lag function is a internal function to adjust to the step value.
}
