% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_variables.R
\name{unique_variables}
\alias{unique_variables}
\title{Unique variables}
\usage{
unique_variables(data)
}
\arguments{
\item{data}{data}
}
\value{
list of variables
}
\description{
Function to generate the column names to write to the file system.
}
