% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_final.R
\name{write_to}
\alias{write_to}
\title{write_to}
\usage{
write_to(obj, write_file = NULL, most_recent)
}
\arguments{
\item{obj}{wizard object}

\item{write_file}{File location to write file}

\item{most_recent}{To include only the most recent lagged features}
}
\value{
wizard object
}
\description{
The function to write or spread the temporal data frame.
}
