% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_final.R
\name{add_diff_predictors}
\alias{add_diff_predictors}
\title{add_diff_predictors}
\usage{
add_diff_predictors(obj, categories = list())
}
\description{
Function to generate iterated difference between lagged features
}
\details{
@param obj wizard object
 @param categories List of categories to generate the iterated difference feature
 @return wizard object
 @export
}
