% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_final.R
\name{add_prop_predictors}
\alias{add_prop_predictors}
\title{add_prop_predictors}
\usage{
add_prop_predictors(obj, categories = list())
}
\description{
Function to generate proportional variation between lagged features
}
\details{
@param obj wizard object
 @param categories List of categories to generate teh proportional variation feature
 @return wizard object
 @export
}
