% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_feature.R
\name{diff_feature}
\alias{diff_feature}
\title{Difference features}
\usage{
diff_feature(.x, .y, window_size, step, lag_compute)
}
\arguments{
\item{window_size}{window size}

\item{final_frame}{final df}
}
\value{
lag frame
}
\description{
The function to compute the difference features.
}
